Remove sorting links
Sorting via links that reload the whole page has gotten impractical
since the addition of local area vs. wider area tabs. A client-side way
to sort tables would be a good replacement, but for the moment I haven't
found any third-party libraries that a) I like and b) work. I might
develop a client-side table sorter in the near future for this.