Revise how response email address is found
CAN'T assume it's the last question in first section of response
Because it's been moved to the second section in SOME surveys

CAN'T identify it by survey details -> ID of question labeled "Email"
Because when Rushville's survey's structure was changed (putting the
email field in the second section), an API call to get the survey
structure still returned the old structure (with email in the first
section).

CAN iterate through all response data looking for a valid email address
More processor-expensive, but it appears to be the only way.