Hide error list and display toggle button
A really long error list that permanently lives on the client page, even
after the community has moved on to another step, would be really
disruptive to the layout. Practically speaking, these errors aren't a
huge deal and many might just be ignored forever by the client, and
that's fine.