Remove returnPath values
This was causing some email authentication errors. SPF validation checks the return-path to make sure that domain matches up with the IP address that's attempting to send the email. This means that for most cases, bsu.edu was sending emails with non-BSU return-path values, and for any non-BSU domains that have SPF validation set up, these emails fail that validation.