Fix error regarding collecting Q&A IDs
A recent change accidentally made it so that Q&A IDs can only be looked up for surveys with records in the database, since the survey's type was being looked up in the database for the purpose of determining whether or not "aware of plan" Q&A IDs should be expected. This logic was moved client-side, where the (proposed) survey record's type is available.