Fix problem reading checkSurveyAssignment data
Previously, the JSON object was what only the contents of the PHP variable $community, but now other variables used by the site layout have been added to that object and 'community' is one of its properties. This fix allows the previous format of the data (which we should return to) to work, as well as the current format.