Prevent $this->request->data from populating form
Kind of a hack. Form field values are set to the appropriate value from $invitees or null, but when they're null, the form still pulls from $this->request->data to populate them. This results in not being able to display a blank form after submitting it with no errors. With this change, the form is looking in $this->request->query (which should always be empty) for values to populate empty fields with.